commit before going to school
git-svn-id: https://subversion.assembla.com/svn/danlatimer/trunk/cs6345_spatial_data_structures/final_project@112 cf56831e-89ec-47d9-9321-222c2ec06555

+++ final_project/PrintManager.cpp
@@ -244,7 +244,7 @@ void PrintManager::PrintPolygon(const std::vector<dnl::Point> &points, Colour *c
 	GraphicsPath path;
 
 	path.StartFigure();
-	for(unsigned int i = 0; i < points.size() - 1; i++)
+	for(int i = 0; i < ((int)points.size()) - 1; i++)
 	{
 		path.AddLine(
 			transformX(points[i].m_x),
